+++ umsgpack.h
@@ -119,12 +119,12 @@ int umsgpack_pack_float(struct umsgpack_packer_buf *buf, float val);
 #if 0
 int umsgpack_pack_double(struct umsgpack_packer_buf *buf, double val);
 #endif
-int umsgpack_pack_map(struct umsgpack_packer_buf *buf, unsigned int num_objects);
-int umsgpack_pack_str(struct umsgpack_packer_buf *buf, char* s, int length);
+int umsgpack_pack_map(struct umsgpack_packer_buf *buf, uint32_t num_objects);
+int umsgpack_pack_str(struct umsgpack_packer_buf *buf, char* s, uint32_t length);
 int umsgpack_pack_bool(struct umsgpack_packer_buf *buf, int val);
 int umsgpack_pack_nil(struct umsgpack_packer_buf *buf);
-void umsgpack_packer_init(struct umsgpack_packer_buf *buf, int size);
-struct umsgpack_packer_buf *umsgpack_alloc(int size);
+void umsgpack_packer_init(struct umsgpack_packer_buf *buf, size_t size);
+struct umsgpack_packer_buf *umsgpack_alloc(size_t size);
 int umsgpack_free(struct umsgpack_packer_buf *buf);
 
 #endif /* UMSGPACK_H_ */
